% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R, R/ex4.R
\name{afisare_repartitie}
\alias{afisare_repartitie}
\title{afisare funtie de repartitie}
\usage{
afisare_repartitie(f, minim, maxim)

afisare_repartitie(f, minim, maxim)
}
\description{
afisare funtie de repartitie

afisare funtie de repartitie
}
