% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R, R/ex6.R
\name{calcul_medie_dispersie_repartitie}
\alias{calcul_medie_dispersie_repartitie}
\title{functie care returneaza media si densitatea in functie de g(x) folosind repartitii cunoscute
g functie introdusa de utilizator
repartitie - repartitia folosita
a, b, n, miu, lambda, sigma - paramterii folositi in functie de repartitia aleasa
lognormal weibull normal chipatrat students exponential beta gamma uniform}
\usage{
calcul_medie_dispersie_repartitie(
  g,
  repartitie = "",
  a = 0,
  b = 0,
  n = 0,
  lambda = 0,
  sigma = 0,
  miu = 0
)

calcul_medie_dispersie_repartitie(
  g,
  repartitie = "",
  a = 0,
  b = 0,
  n = 0,
  lambda = 0,
  sigma = 0,
  miu = 0
)
}
\description{
functie care returneaza media si densitatea in functie de g(x) folosind repartitii cunoscute
g functie introdusa de utilizator
repartitie - repartitia folosita
a, b, n, miu, lambda, sigma - paramterii folositi in functie de repartitia aleasa
lognormal weibull normal chipatrat students exponential beta gamma uniform

functie care returneaza media si densitatea in functie de g(x) folosind repartitii cunoscute
g functie introdusa de utilizator
repartitie - repartitia folosita
a, b, n, miu, lambda, sigma - paramterii folositi in functie de repartitia aleasa
lognormal weibull normal chipatrat students exponential beta gamma uniform
}
