% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R, R/ex12.R
\name{diferenta_convolutie}
\alias{diferenta_convolutie}
\title{functie pentru diferenta
f - functia de densitate de probabilitate a variabilei x
g - ... Y}
\usage{
diferenta_convolutie(f, g)

diferenta_convolutie(f, g)
}
\description{
functie pentru diferenta
f - functia de densitate de probabilitate a variabilei x
g - ... Y

functie pentru diferenta
f - functia de densitate de probabilitate a variabilei x
g - ... Y
}
