% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ex2}
\alias{ex2}
\title{daca eroare inseamna ca integrala nu merge
=> nu e o integrala finita
=> ori functia nu e pozitiva ori integrala nu e finita
=> ori orice alta eroare poate sa dea integrala
return (FALSE)}
\usage{
ex2(FUN)
}
\description{
daca eroare inseamna ca integrala nu merge
=> nu e o integrala finita
=> ori functia nu e pozitiva ori integrala nu e finita
=> ori orice alta eroare poate sa dea integrala
return (FALSE)
}
