% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R, R/ex6.R
\name{medie_dispersie}
\alias{medie_dispersie}
\title{functie ce calculeaza media si dispersia
g - functie introdusa de utilizator
d - functie de densitate
minim - capatul de jos al integralei
maxim - capatul de sus al integralei
paste - concatenare dupa conversia la char}
\usage{
medie_dispersie(g, d, minim, maxim)

medie_dispersie(g, d, minim, maxim)
}
\description{
functie ce calculeaza media si dispersia
g - functie introdusa de utilizator
d - functie de densitate
minim - capatul de jos al integralei
maxim - capatul de sus al integralei
paste - concatenare dupa conversia la char

functie ce calculeaza media si dispersia
g - functie introdusa de utilizator
d - functie de densitate
minim - capatul de jos al integralei
maxim - capatul de sus al integralei
paste - concatenare dupa conversia la char
}
