% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{momentOrdinK}
\alias{momentOrdinK}
\title{daca eroare inseamna ca integrala nu merge
=> nu e o integrala finita
=> ori functia nu e pozitiva ori integrala nu e finita
=> ori orice alta eroare poate sa dea integrala
return (FALSE)}
\usage{
momentOrdinK(FUN, k)
}
\description{
daca eroare inseamna ca integrala nu merge
=> nu e o integrala finita
=> ori functia nu e pozitiva ori integrala nu e finita
=> ori orice alta eroare poate sa dea integrala
return (FALSE)
}
