% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot_densitate}
\alias{plot_densitate}
\title{afisare functia de densitatea a  probabilitatii pentru variabile aleatoare
s - cat de apropiate sunt punctele
trebuie}
\usage{
plot_densitate(f, minim, maxim)
}
\description{
afisare functia de densitatea a  probabilitatii pentru variabile aleatoare
s - cat de apropiate sunt punctele
trebuie
}
