% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{suma_convolutie}
\alias{suma_convolutie}
\title{suma si diferenta a doua variabile aleatoare continue
diferite folosind formula de convolutie
functie ce calculeaza suma
f - functia de densitate de probabilitate a variabilei x
g - ... Y}
\usage{
suma_convolutie(f, g)
}
\description{
suma si diferenta a doua variabile aleatoare continue
diferite folosind formula de convolutie
functie ce calculeaza suma
f - functia de densitate de probabilitate a variabilei x
g - ... Y
}
